pyproject.toml
[project]
name = "cold-mail-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.19.0",
]
